Del: Delete KnotDataManager
KnotDataManager 관련 코드 및 파일 삭제
